+++ example/client.js
@@ -1,8 +1,0 @@
-import { createOpenAPI } from '@tencent-connect/bot-node-sdk';
-
-const testConfig = {
-  appID: '',
-  token: '',
-};
-
-export const client = createOpenAPI(testConfig);