+++ scripts/dev.js
@@ -2,5 +2,5 @@ const { execSync } = require('child_process');
 
 try {
   execSync('npm link');
-  execSync('cd example && npm link @tencent-connect/bot-node-sdk');
+  execSync('cd example && npm link qq-guild-bot');
 } catch (error) {}