+++ src/bot.ts
@@ -1,15 +1,11 @@
 import { apiVersion, OpenAPI, v1Setup } from '@src/openapi/v1/openapi';
 import { versionMapping } from '@src/openapi/openapi';
 import { APIVersion, Config } from './types/openapi';
-import { GetWsParam } from './types/websocket-types';
+import { GetWsParam } from '@src/types';
 import WebsocketClient from './client/client';
 
-function init() {
-  // 注册v1接口
-  v1Setup();
-}
-// 进行初始化
-init();
+// 注册v1接口
+v1Setup();
 
 let defaultImpl = versionMapping[apiVersion] as typeof OpenAPI;