+++ src/openapi/v1/openapi.ts
@@ -1,17 +1,5 @@
+/* eslint-disable prefer-promise-reject-errors */
 import { register } from '@src/openapi/openapi';
-import {
-  AudioAPI,
-  ChannelAPI,
-  ChannelPermissionsAPI,
-  Config,
-  DirectMessageAPI,
-  GuildAPI,
-  IOpenAPI,
-  MeAPI,
-  MemberAPI,
-  MessageAPI,
-  RoleAPI,
-} from '@src/types/openapi';
 import resty, { RequestOptions, RestyResponse } from 'resty-client';
 import Guild from './guild';
 import Channel from './channel';
@@ -22,13 +10,34 @@ import Role from './role';
 import DirectMessage from './direct-message';
 import ChannelPermissions from './channel-permissions';
 import Audio from './audio';
+import Mute from './mute';
+import Announce from './announce';
+import Schedule from './schedule';
+import GuildPermissions from './guiild-permissions';
+import { addUserAgent, addAuthorization, buildUrl } from '@src/utils/utils';
+import {
+  GuildAPI,
+  ChannelAPI,
+  MeAPI,
+  MessageAPI,
+  Config,
+  IOpenAPI,
+  MemberAPI,
+  RoleAPI,
+  DirectMessageAPI,
+  ChannelPermissionsAPI,
+  AudioAPI,
+  MuteAPI,
+  ScheduleAPI,
+  AnnounceAPI,
+  GuildPermissionsAPI,
+} from '@src/types';
 
 export const apiVersion = 'v1';
 
 export class OpenAPI implements IOpenAPI {
   static newClient(config: Config) {
-    const client = new OpenAPI(config);
-    return client;
+    return new OpenAPI(config);
   }
 
   config: Config = {
@@ -41,13 +50,19 @@ export class OpenAPI implements IOpenAPI {
   public messageApi!: MessageAPI;
   public memberApi!: MemberAPI;
   public roleApi!: RoleAPI;
+  public muteApi!: MuteAPI;
+  public announceApi!: AnnounceAPI;
+  public scheduleApi!: ScheduleAPI;
   public directMessageApi!: DirectMessageAPI;
   public channelPermissionsApi!: ChannelPermissionsAPI;
   public audioApi!: AudioAPI;
+  public guildPermissionsApi!: GuildPermissionsAPI;
+
   constructor(config: Config) {
     this.config = config;
     this.register(this);
   }
+
   public register(client: IOpenAPI) {
     // 注册聚合client
     client.guildApi = new Guild(this.request, this.config);
@@ -56,27 +71,50 @@ export class OpenAPI implements IOpenAPI {
     client.messageApi = new Message(this.request, this.config);
     client.memberApi = new Member(this.request, this.config);
     client.roleApi = new Role(this.request, this.config);
+    client.muteApi = new Mute(this.request, this.config);
+    client.announceApi = new Announce(this.request, this.config);
+    client.scheduleApi = new Schedule(this.request, this.config);
     client.directMessageApi = new DirectMessage(this.request, this.config);
     client.channelPermissionsApi = new ChannelPermissions(this.request, this.config);
     client.audioApi = new Audio(this.request, this.config);
+    client.guildPermissionsApi = new GuildPermissions(this.request, this.config);
   }
   // 基础rest请求
   public request<T extends Record<any, any> = any>(options: RequestOptions): Promise<RestyResponse<T>> {
     const { appID, token } = this.config;
-    options.headers = {
-      ...options.headers,
-      'User-Agent': apiVersion,
-      Authorization: `Bot ${appID}.${token}`,
-    };
 
-    // 简化错误信息
+    options.headers = { ...options.headers };
+
+    // 添加 UA
+    addUserAgent(options.headers);
+    // 添加鉴权信息
+    addAuthorization(options.headers, appID, token);
+    // 组装完整Url
+    const botUrl = buildUrl(options.url, this.config.sandbox);
+
+    // 简化错误信息，后续可考虑通过中间件形式暴露给用户自行处理
     resty.useRes(
       (result) => result,
-      (error) => Promise.reject(error.response.data),
+      (error) => {
+        let traceid = error?.response?.headers?.['x-tps-trace-id'];
+        if (error?.response?.data) {
+          return Promise.reject({
+            ...error.response.data,
+            traceid,
+          });
+        }
+        if (error?.response) {
+          return Promise.reject({
+            ...error.response,
+            traceid,
+          });
+        }
+        return Promise.reject(error);
+      },
     );
 
     const client = resty.create(options);
-    return client.request<T>(options.url!, options);
+    return client.request<T>(botUrl!, options);
   }
 }