+++ src/types/websocket-types.ts
@@ -1,5 +1,7 @@
 import { apiVersion } from '@src/openapi/v1/openapi';
 import { getURL } from '@src/openapi/v1/resource';
+import { buildUrl } from '@src/utils/utils';
+
 // websocket建立成功回包
 export interface wsResData {
   op: number; // opcode ws的类型
@@ -27,8 +29,10 @@ export interface EventTypes {
 export interface GetWsParam {
   appID: string;
   token: string;
+  sandbox?: boolean;
   shards?: Array<number>;
-  intents?: Array<string>;
+  intents?: Array<AvailableIntentsEventsEnum>;
+  maxRetry?: number;
 }
 
 // 请求ws地址回包对象
@@ -66,23 +70,64 @@ export enum OpCode {
   HEARTBEAT_ACK = 11, // 当发送心跳成功之后，就会收到该消息
 }
 
+// 可使用的intents事件类型
+export enum AvailableIntentsEventsEnum {
+  GUILDS = 'GUILDS',
+  GUILD_MEMBERS = 'GUILD_MEMBERS',
+  GUILD_MESSAGES = 'GUILD_MESSAGES',
+  GUILD_MESSAGE_REACTIONS = 'GUILD_MESSAGE_REACTIONS',
+  DIRECT_MESSAGE = 'DIRECT_MESSAGE',
+  FORUM_EVENT = 'FORUM_EVENT',
+  AUDIO_ACTION = 'AUDIO_ACTION',
+  AT_MESSAGES = 'AT_MESSAGES',
+  MESSAGE_AUDIT = 'MESSAGE_AUDIT',
+}
+
 // OpenAPI传过来的事件类型
-export const WsEventType: { [key: string]: string } = {
-  GUILD_CREATE: 'GUILDS', // 频道创建
-  GUILD_UPDATE: 'GUILDS', // 频道更新
-  GUILD_DELETE: 'GUILDS', // 频道删除
-  CHANNEL_CREATE: 'GUILDS', // 子频道创建
-  CHANNEL_UPDATE: 'GUILDS', // 子频道更新
-  CHANNEL_DELETE: 'GUILDS', // 子频道删除
-  GUILD_MEMBER_ADD: 'GUILD_MEMBERS', // 频道成员加入
-  GUILD_MEMBER_UPDATE: 'GUILD_MEMBERS', // 频道成员更新
-  GUILD_MEMBER_REMOVE: 'GUILD_MEMBERS', // 频道成员移除
-  DIRECT_MESSAGE_CREATE: 'DIRECT_MESSAGE', // 当收到用户发给机器人的私信消息时
-  AT_MESSAGE_CREATE: 'AT_MESSAGES', // 机器人被@时触发
-  AUDIO_START: 'AUDIO_ACTION', // 音频开始播放
-  AUDIO_FINISH: 'AUDIO_ACTION', // 音频结束播放
-  AUDIO_ON_MIC: 'AUDIO_ACTION', // 机器人上麦
-  AUDIO_OFF_MIC: 'AUDIO_ACTION', // 机器人下麦
+export const WsEventType: { [key: string]: AvailableIntentsEventsEnum } = {
+  //  ======= GUILDS ======
+  GUILD_CREATE: AvailableIntentsEventsEnum.GUILDS, // 频道创建
+  GUILD_UPDATE: AvailableIntentsEventsEnum.GUILDS, // 频道更新
+  GUILD_DELETE: AvailableIntentsEventsEnum.GUILDS, // 频道删除
+  CHANNEL_CREATE: AvailableIntentsEventsEnum.GUILDS, // 子频道创建
+  CHANNEL_UPDATE: AvailableIntentsEventsEnum.GUILDS, // 子频道更新
+  CHANNEL_DELETE: AvailableIntentsEventsEnum.GUILDS, // 子频道删除
+
+  //  ======= GUILD_MEMBERS ======
+  GUILD_MEMBER_ADD: AvailableIntentsEventsEnum.GUILD_MEMBERS, // 频道成员加入
+  GUILD_MEMBER_UPDATE: AvailableIntentsEventsEnum.GUILD_MEMBERS, // 频道成员更新
+  GUILD_MEMBER_REMOVE: AvailableIntentsEventsEnum.GUILD_MEMBERS, // 频道成员移除
+
+  //  ======= AUDIO_ACTION ======
+  AUDIO_START: AvailableIntentsEventsEnum.AUDIO_ACTION, // 音频开始播放
+  AUDIO_FINISH: AvailableIntentsEventsEnum.AUDIO_ACTION, // 音频结束播放
+  AUDIO_ON_MIC: AvailableIntentsEventsEnum.AUDIO_ACTION, // 机器人上麦
+  AUDIO_OFF_MIC: AvailableIntentsEventsEnum.AUDIO_ACTION, // 机器人下麦
+
+  //  ======= GUILD_MESSAGE_REACTIONS ======
+  MESSAGE_REACTION_ADD: AvailableIntentsEventsEnum.GUILD_MESSAGE_REACTIONS, // 为消息添加表情表态
+  MESSAGE_REACTION_REMOVE: AvailableIntentsEventsEnum.GUILD_MESSAGE_REACTIONS, // 为消息删除表情表态
+
+  //  ======= MESSAGE_AUDIT ======
+  MESSAGE_AUDIT_PASS: AvailableIntentsEventsEnum.MESSAGE_AUDIT, // / 消息审核通过
+  MESSAGE_AUDIT_REJECT: AvailableIntentsEventsEnum.MESSAGE_AUDIT, // / 消息审核不通过
+
+  //  ======= FORUM_EVENT ======
+  THREAD_CREATE: AvailableIntentsEventsEnum.FORUM_EVENT, // / 当用户创建主题时
+  THREAD_UPDATE: AvailableIntentsEventsEnum.FORUM_EVENT, // 当用户删除主题时
+  POST_CREATE: AvailableIntentsEventsEnum.FORUM_EVENT, // 当用户创建帖子时
+  POST_DELETE: AvailableIntentsEventsEnum.FORUM_EVENT, // 当用户删除帖子时
+  REPLY_CREATE: AvailableIntentsEventsEnum.FORUM_EVENT, // 当用户回复评论时
+  REPLY_DELETE: AvailableIntentsEventsEnum.FORUM_EVENT, // 当用户回复评论时
+
+  //  ======= DIRECT_MESSAGE ======
+  DIRECT_MESSAGE_CREATE: AvailableIntentsEventsEnum.DIRECT_MESSAGE, // 当收到用户发给机器人的私信消息时
+
+  //  ======= AT_MESSAGES ======
+  AT_MESSAGE_CREATE: AvailableIntentsEventsEnum.AT_MESSAGES, // 机器人被@时触发
+
+  //  ======= GUILD_MESSAGES ======
+  MESSAGE_CREATE: AvailableIntentsEventsEnum.GUILD_MESSAGES, // 机器人收到频道消息时触发
 };
 
 export const WSCodes = {
@@ -126,10 +171,12 @@ export const WebsocketCloseReason = [
   {
     code: 4008,
     reason: '发送 payload 过快，请重新连接，并遵守连接后返回的频控信息',
+    resume: true,
   },
   {
     code: 4009,
     reason: '连接过期，请重连',
+    resume: true,
   },
   {
     code: 4010,
@@ -155,13 +202,29 @@ export const WebsocketCloseReason = [
     code: 4900,
     reason: '内部错误，请重连',
   },
+  {
+    code: 4914,
+    reason: '机器人已下架,只允许连接沙箱环境,请断开连接,检验当前连接环境',
+  },
+  {
+    code: 4915,
+    reason: '机器人已封禁,不允许连接,请断开连接,申请解封后再连接',
+  },
 ];
 
+export type IntentEventsMapType = {
+  [key in AvailableIntentsEventsEnum]: number;
+};
+
 // 用户输入的intents类型
-export const IntentEvents: { [key: string]: number } = {
+export const IntentEvents: IntentEventsMapType = {
   GUILDS: 1 << 0,
   GUILD_MEMBERS: 1 << 1,
+  GUILD_MESSAGES: 1 << 9,
+  GUILD_MESSAGE_REACTIONS: 1 << 10,
   DIRECT_MESSAGE: 1 << 12,
+  MESSAGE_AUDIT: 1 << 27,
+  FORUM_EVENT: 1 << 28,
   AUDIO_ACTION: 1 << 29,
   AT_MESSAGES: 1 << 30,
 };
@@ -195,12 +258,13 @@ export const SessionEvents = {
   DISCONNECT: 'DISCONNECT', // 断线
   EVENT_WS: 'EVENT_WS', // 内部通信
   RESUMED: 'RESUMED', // 重连
+  DEAD: 'DEAD', // 连接已死亡，请检查网络或重启
 };
 
 // ws地址配置
-export const WsObjRequestOptions = {
+export const WsObjRequestOptions = (sandbox: boolean) => ({
   method: 'GET' as const,
-  url: getURL('wsInfo'),
+  url: buildUrl(getURL('wsInfo'), sandbox),
   headers: {
     Accept: '*/*',
     'Accept-Encoding': 'utf-8',
@@ -209,4 +273,4 @@ export const WsObjRequestOptions = {
     'User-Agent': apiVersion,
     Authorization: '',
   },
-};
+});